Gathering the already-done/approved music.
The numbered files are for use in dungeons, and will be cycled through when the player has traversed a certain number of rooms, with more to come (12 in all).

'Caffeine' is for when the player is invincible, 'Chara Select' is for the character customization and how-to-play screens, the rest all do what their names imply.

They'll need to be converted to .ogg before use, since Godot cannot cope with XM files (without a plugin).